Reformat .travis.yml, 80 characters rule
Use multiline yml strings to break newlines.  Newlines in yaml
multiline strings are replaced with a space character, so no
backslashes are requiered.

For some reason unknown to me at this time it seems to like travis
decides to not replace the newline character after "sudo ln -s" with a
space.  This is why we have to keep some backslashes in before_install
section.