Refactor stage1.py for better maintainability
This commit seperates calls to nix-shell and nix-instantiate into a seperate class.
Unfortunatly the behavior of pypi2nix is changed slightly now.  It used to be that
the user had to escape $-characters in the --extra-env argument as it was internally
passed to a shell instance.  Now this is not necessary anymore.