Use colon as a delimiter for splitting NIX_PATH
Having a NIX_PATH with multiple values causes pypi2nix to pass it to Nix
via *one* "-I" argument, which is not what Nix expects, thus resulting
in errors like this:

  warning: Nix search path entry '...' does not exist, ignoring

By default click splits environment variables on whitespace[1], rather
than colon, which is the separator for NIX_PATH.

Splitting on colon is only done via the File and Path types in click,
but they also either open the file or do checks whether the path exists.

These types are fine for things such as PATH or LD_LIBRARY_PATH and the
likes, but NIX_PATH has a different format in that it allows bare paths
and assignments (like eg. "nixpkgs=/some/path").

So in order to fix this, I just added a new type that inherits from
StringParamType, but correctly splits on colons.

[1]: https://click.palletsprojects.com/en/7.x/options/#multiple-values-from-environment-values

Signed-off-by: aszlig <aszlig@nix.build>