cmd: Don't default to merging stdout with stderr
This is somewhat related to #280 (Use colon as a delimiter for splitting
NIX_PATH), because since the NIX_PATH got splitted in the wrong way, a
warning was emitted to stdout.

So when getting a JSON serialized output in Pip.default_environment(),
the output from the Nix (shell) command contained the warning at the
top, which in turn results in a JSONDecodeError.

I think the default for this function should be to *not* merge stderr,
which will eventually be printed to the TTY where the user can
immediately see the warnings.

Signed-off-by: aszlig <aszlig@nix.build>