cli: Do not pick up NIX_PATH as default nix_path
By default Nix merges the -I options with the values of NIX_PATH, which
is exactly what is even documented for the -I/--nix-path option in
pypi2nix, so we shouldn't actually base this option on the NIX_PATH
environment variable.

This is a follow-up to #280, where I initially wanted to fix path
splitting (so it's using colon instead of whitespace), but it's a bit
more complicated, because NIX_PATH is splitted on colons but may also
contain URLs with a colon.

Signed-off-by: aszlig <aszlig@nix.build>
Closes: #280