prefetch_hg: Capture stderr as well
It seems that the hash and hg revisions are emitted to stderr instead of
stdout, so when changing the defaults to not capture stderr, the
prefetch_hg() function thus is no longer able to pick up the right
revision.

Simply re-adding stderr=subprocess.STDOUT gives prefetch_hg() the same
behaviour as before but without capturing stderr by default.

Signed-off-by: aszlig <aszlig@nix.build>