prefetch_url: Fix getting output hash
Since prefetch_url() relied previously on stderr being merged into
stdout, the change to not merge stdout and stderr now causes
prefetch_url() to no longer work because it expects three(!) lines.

The latter is very fragile, because it means that we're heavily relying
on a Nix implementation detail on how progress is reported. After all,
we're only interested about what is actually returned via stdout.

To make sure we don't get a bunch of status output from
nix-prefetch-url, I redirected stderr to /dev/null, so that we do not
get flooded with status information while running pypi2nix
non-verbosely.

One downside of this is, that we now no longer get an error message on
*why* a download has failed. However, we then still know the
nix-prefetch-url command that was executed and can simply replicate it
manually. I know it's not very nice, but at least it's less error-prone
than before. A better solution would be to capture stdout and stderr
separately and emit stderr if the command didn't return successfully.

Signed-off-by: aszlig <aszlig@nix.build>