utils: Fix duplicate newlines inserted by cmd()
Ideally we want to have the exact output that we get from the command
we're running, but using readline() on stdout of the process will *not*
strip the newlines, so when we join them again using newlines, we
suddenly end up with twice the amount of newlines.

I also added a small unit test case to make sure this won't break again
in the future.

Signed-off-by: aszlig <aszlig@nix.build>