pip.freeze: Add newline at the end of file
Not properly terminating the file is not only creating errors with Git
and various editors but also will make sure that you'll get ugly output
when using "cat" on the file.

If the file is empty however, we don't want to have any newline
appended, so a condition is there to make sure it doesn't get appended
in this case.

Signed-off-by: aszlig <aszlig@nix.build>