cmd: Don't default to merging stdout with stderr (#281)
* cmd: Don't default to merging stdout with stderr

This is somewhat related to #280 (Use colon as a delimiter for splitting
NIX_PATH), because since the NIX_PATH got splitted in the wrong way, a
warning was emitted to stdout.

So when getting a JSON serialized output in Pip.default_environment(),
the output from the Nix (shell) command contained the warning at the
top, which in turn results in a JSONDecodeError.

I think the default for this function should be to *not* merge stderr,
which will eventually be printed to the TTY where the user can
immediately see the warnings.

Signed-off-by: aszlig <aszlig@nix.build>

* prefetch_hg: Capture stderr as well

It seems that the hash and hg revisions are emitted to stderr instead of
stdout, so when changing the defaults to not capture stderr, the
prefetch_hg() function thus is no longer able to pick up the right
revision.

Simply re-adding stderr=subprocess.STDOUT gives prefetch_hg() the same
behaviour as before but without capturing stderr by default.

Signed-off-by: aszlig <aszlig@nix.build>

* prefetch_url: Fix getting output hash

Since prefetch_url() relied previously on stderr being merged into
stdout, the change to not merge stdout and stderr now causes
prefetch_url() to no longer work because it expects three(!) lines.

The latter is very fragile, because it means that we're heavily relying
on a Nix implementation detail on how progress is reported. After all,
we're only interested about what is actually returned via stdout.

To make sure we don't get a bunch of status output from
nix-prefetch-url, I redirected stderr to /dev/null, so that we do not
get flooded with status information while running pypi2nix
non-verbosely.

One downside of this is, that we now no longer get an error message on
*why* a download has failed. However, we then still know the
nix-prefetch-url command that was executed and can simply replicate it
manually. I know it's not very nice, but at least it's less error-prone
than before. A better solution would be to capture stdout and stderr
separately and emit stderr if the command didn't return successfully.

Signed-off-by: aszlig <aszlig@nix.build>